Add Jackson dependency for JSON processing
