+++ build.gradle
@@ -9,6 +9,7 @@ repositories {
 dependencies {
     implementation 'junit:junit:4.13.1'
     testImplementation('org.junit.jupiter:junit-jupiter:5.6.0')
+    implementation 'com.fasterxml.jackson.core:jackson-databind:2.12.5'
 }
 
 test {
