--- src/main/java/game_ui/GameBoardView.java
@@ -1,4 +1,0 @@
-package game_ui;
-
-public class GameBoardView {
-}
