--- src/main/java/interface_adapters/GamePrepFailed.java
@@ -1,8 +1,0 @@
-package interface_adapters;
-
-public class GamePrepFailed extends RuntimeException {
-
-    public GamePrepFailed(String message) {
-        super(message);
-    }
-}
