+++ src/main/java/interface_adapters/view_models/SetupScreenModel.java
@@ -1,0 +1,6 @@
+package interface_adapters;
+
+public class SetupScreenModel {
+
+
+}
